+++ src/test/groovy/br/com/unopay/api/market/MarketTemplateLoader.java
@@ -127,7 +127,7 @@ public class MarketTemplateLoader implements TemplateLoader {
         Fixture.of(BonusBilling.class).addTemplate("valid", new Rule(){{
             add("createdDateTime", instant("now"));
             add("payer", one(Person.class, "physical"));
-            add("total", random(Double.class, range(2, 300)));
+            add("total", random(BigDecimal.class, range(2, 300)));
             add("processedAt", instant("now"));
             add("number",  regex("\\w{100}"));
             add("expiration", instant("tomorrow"));
